Fix PWA manifest and Docker BASE_PATH support
When running the web app on a subpath, the scope needs to be set. It
seems that `scope` essentially sets the base URL for the manifest
itself, so this needs to be updated as well as the `start_url` as was
being done before.

The SW registration also needs to explicitly register the `/BASE_PATH/`
scope, instead of the default `/` scope now.